fix flow cycle errors and add comments
